requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.
#
django-rosetta==0.5.6
PIL
psycopg2
BeautifulSoup
celery
django-celery
http://projects.unbit.it/downloads/uwsgi-latest.tar.gz

